refactor(wallet): migrate keyring + storage instances to directory convention
Adopt the per-controller directory + `types.ts` convention introduced for
`approval-controller` in #8953 for the two pre-existing instances, so the
package isn't left in two styles:

- Move `keyring-controller.ts` and `storage-service.ts` into
  `keyring-controller/` and `storage-service/` directories.
- Pull the keyring encryptor out of the controller file into its own
  `keyring-controller/encryptor.ts` sibling (with colocated `encryptor.test.ts`).
- Add per-controller `KeyringControllerInstanceOptions` /
  `StorageServiceInstanceOptions` types in each directory's `types.ts`, wired
  into `InstanceSpecificOptions` (the public option shape is unchanged).
- Update `instances/index.ts` import paths and the `## Initialization`
  CODEOWNERS entries to the directory form (keyring mirrors its package owners;
  add a storage-service entry mirroring its package owners).

Internal refactor only — no public exports or option shapes change.
