+++ packages/wallet/src/initialization/instances/keyring-controller/encryptor.test.ts
@@ -1,38 +1,6 @@
-import { KeyringController } from '@metamask/keyring-controller';
-import type {
-  KeyringControllerMessenger,
-  KeyringControllerOptions,
-} from '@metamask/keyring-controller';
 import { webcrypto } from 'crypto';
 
-import { encryptorFactory, keyringController } from './keyring-controller';
-
-jest.mock('@metamask/keyring-controller', () => ({
-  ...jest.requireActual('@metamask/keyring-controller'),
-  KeyringController: jest.fn(),
-}));
-
-describe('keyringController', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('forwards keyringV2Builders to the KeyringController', () => {
-    const keyringV2Builders = [
-      Object.assign(jest.fn(), { type: 'Test Keyring V2' }),
-    ] as KeyringControllerOptions['keyringV2Builders'];
-
-    keyringController.init({
-      state: undefined,
-      messenger: {} as KeyringControllerMessenger,
-      options: { keyringV2Builders },
-    });
-
-    expect(KeyringController).toHaveBeenCalledWith(
-      expect.objectContaining({ keyringV2Builders }),
-    );
-  });
-});
+import { encryptorFactory } from './encryptor';
 
 describe('encryptorFactory', () => {
   beforeAll(() => {
