+++ packages/wallet/src/initialization/instances/keyring-controller/encryptor.ts
@@ -17,13 +17,6 @@ import {
   generateSalt,
 } from '@metamask/browser-passworder';
 import type { Encryptor } from '@metamask/keyring-controller';
-import {
-  KeyringController,
-  KeyringControllerMessenger,
-} from '@metamask/keyring-controller';
-import { Messenger } from '@metamask/messenger';
-
-import { InitializationConfiguration } from '../types';
 
 /**
  * A factory function for the encrypt method of the browser-passworder library,
@@ -160,25 +153,3 @@ export type GenericEncryptor =
       KeyDerivationOptions,
       MobileEncryptionResult
     >;
-
-export const keyringController: InitializationConfiguration<
-  KeyringController,
-  KeyringControllerMessenger
-> = {
-  name: 'KeyringController',
-  init: ({ state, messenger, options }) =>
-    new KeyringController({
-      state,
-      messenger,
-      keyringBuilders: options.keyringBuilders,
-      keyringV2Builders: options.keyringV2Builders,
-      encryptor: (options.encryptor ?? encryptorFactory(600_000)) as Encryptor<
-        EncryptionKey | CryptoKey
-      >,
-    }),
-  getMessenger: (parent) =>
-    new Messenger({
-      namespace: 'KeyringController',
-      parent,
-    }),
-};
